+++ python/cdshealpix/nested/healpix.py
@@ -364,8 +364,8 @@ def vertices(ipix, depth, step=1, num_threads=0):
     ipix = ipix.astype(np.uint64)
 
     # Allocation of the array containing the resulting coordinates
-    lon = np.zeros(ipix.shape + (4 * step,))
-    lat = np.zeros(ipix.shape + (4 * step,))
+    lon = np.zeros((*ipix.shape, 4 * step))
+    lat = np.zeros((*ipix.shape, 4 * step))
     num_threads = np.uint16(num_threads)
 
     cdshealpix.vertices(depth, ipix, step, lon, lat, num_threads)
@@ -465,7 +465,7 @@ def neighbours(ipix, depth, num_threads=0):
     ipix = ipix.astype(np.uint64)
 
     # Allocation of the array containing the neighbours
-    neighbours = np.zeros(ipix.shape + (9,), dtype=np.int64)
+    neighbours = np.zeros((*ipix.shape, 9), dtype=np.int64)
     num_threads = np.uint16(num_threads)
     cdshealpix.neighbours(depth, ipix, neighbours, num_threads)
 
@@ -511,8 +511,8 @@ def external_neighbours(ipix, depth, delta_depth, num_threads=0):
 
     # Allocation of the array containing the neighbours
     num_external_cells_on_edges = 4 << delta_depth
-    edge_cells = np.zeros(ipix.shape + (num_external_cells_on_edges,), dtype=np.uint64)
-    corner_cells = np.zeros(ipix.shape + (4,), dtype=np.int64)
+    edge_cells = np.zeros((*ipix.shape, num_external_cells_on_edges), dtype=np.uint64)
+    corner_cells = np.zeros((*ipix.shape, 4), dtype=np.int64)
 
     num_threads = np.uint16(num_threads)
     cdshealpix.external_neighbours(
@@ -665,13 +665,13 @@ def polygon_search(lon, lat, depth, flat=False):
     num_vertices = lon.shape[0]
 
     if num_vertices < 3:
-        raise IndexError("There must be at least 3 vertices in order to form a polygon")
+        raise ValueError("There must be at least 3 vertices in order to form a polygon")
 
     # Check that there is at least 3 distinct vertices.
     vertices = np.vstack((lon, lat)).T
     distinct_vertices = np.unique(vertices, axis=0)
     if distinct_vertices.shape[0] < 3:
-        raise IndexError(
+        raise ValueError(
             "There must be at least 3 distinct vertices in order to form a polygon"
         )
 
@@ -1004,35 +1004,32 @@ def bilinear_interpolation(lon, lat, depth, num_threads=0):
     if not (isinstance(lat, Latitude)):
         raise ValueError("`lat` must be of type `astropy.coordinates.Latitude`")
 
-    # We could have continued to use `.to_value(u.rad)` instead of `.rad`.
-    # Although `to_value` is more generical (method of Quantity),
-    # Longitude/Latitude ensure that the values given to the contain are in the correct ranges.
-    lon = np.atleast_1d(lon.rad)
-    lat = np.atleast_1d(lat.rad)
-
     if depth < 0 or depth > 29:
         raise ValueError("Depth must be in the [0, 29] closed range")
+
+    lon = np.atleast_1d(lon.rad)
+    lat = np.atleast_1d(lat.rad)
 
     if lon.shape != lat.shape:
         raise ValueError(
             "The number of longitudes does not match with the number of latitudes given"
         )
 
-    # Useless since we test isinstance at the beginning of the function
-    # if ((lat < np.pi/2.0) | (lat > np.pi/2.0)).any():
-    #    raise ValueError("Lat must be in [-pi/2, 2pi/2]")
-
     num_coords = lon.shape
 
     # Retrieve nan and infinite values
-    mask_lon_invalid = np.isnan(lon) | ~np.isfinite(lon)
-    mask_lat_invalid = np.isnan(lat) | ~np.isfinite(lat)
+    mask_lon_invalid = np.ma.masked_invalid(lon).mask
+    mask_lat_invalid = np.ma.masked_invalid(lat).mask
+
     mask_invalid = mask_lon_invalid | mask_lat_invalid
+
+    lon = np.ma.fix_invalid(lon, mask_invalid, fill_value=0 * u.deg)
+    lat = np.ma.fix_invalid(lat, mask_invalid, fill_value=0 * u.deg)
 
     mask_invalid = np.repeat(mask_invalid[:, np.newaxis], 4, axis=mask_invalid.ndim)
 
-    ipix = np.empty(shape=num_coords + (4,), dtype=np.uint64)
-    weights = np.empty(shape=num_coords + (4,), dtype=np.float64)
+    ipix = np.empty(shape=(*num_coords, 4), dtype=np.uint64)
+    weights = np.empty(shape=(*num_coords, 4), dtype=np.float64)
 
     num_threads = np.uint16(num_threads)